+++ libsrc/leddevice/CMakeLists.txt
@@ -40,7 +40,6 @@ SET(Leddevice_HEADERS
 	${CURRENT_SOURCE_DIR}/LedDeviceHyperionUsbasp.h
 	${CURRENT_SOURCE_DIR}/LedDeviceTpm2.h
 	${CURRENT_SOURCE_DIR}/LedDeviceTpm2net.h
-	${CURRENT_SOURCE_DIR}/LedDeviceTpm2netnew.h
 	${CURRENT_SOURCE_DIR}/LedDeviceAtmo.h
 )
 
@@ -69,7 +68,6 @@ SET(Leddevice_SOURCES
 	${CURRENT_SOURCE_DIR}/LedDevicePhilipsHue.cpp
 	${CURRENT_SOURCE_DIR}/LedDeviceTpm2.cpp
 	${CURRENT_SOURCE_DIR}/LedDeviceTpm2net.cpp
-	${CURRENT_SOURCE_DIR}/LedDeviceTpm2netnew.cpp
 	${CURRENT_SOURCE_DIR}/LedDeviceAtmo.cpp
 )
