--- libsrc/leddevice/schemas/schema-tpm2netnew.json
@@ -1,7 +1,0 @@
-{
-	"type":"object",
-	"required":true,
-	"properties":{
-	},
-	"additionalProperties": true
-}
